sensor-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
kafka-python>=2.0.2
requests>=2.31.0
python-dateutil>=2.8.2
jinja2>=3.1.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
gunicorn>=21.2.0
